# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "project akhir",
    "description": "",
    "contact": {}
  },
  "host": "localhost:8000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/auth/relogin": {
      "post": {
        "description": "",
        "summary": "Relogin",
        "tags": [
          "auth"
        ],
        "operationId": "Relogin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/seller/product/54": {
      "get": {
        "description": "",
        "summary": "Get ID Product",
        "tags": [
          "seller"
        ],
        "operationId": "GetIDProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/seller/product/56": {
      "delete": {
        "description": "",
        "summary": "Delete Product",
        "tags": [
          "seller"
        ],
        "operationId": "DeleteProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
...
}